devided the code in multiples files
