+++ config/sources_bonus.mk
@@ -6,7 +6,7 @@
 #    By: brappo <brappo@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/06/19 11:26:46 by mhotting          #+#    #+#              #
-#    Updated: 2024/06/26 13:54:50 by brappo           ###   ########.fr        #
+#    Updated: 2024/06/26 14:32:04 by brappo           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -122,8 +122,9 @@ PATHFINDING_FILES_BONUS			=	add_path_node_bonus.c		\
 									get_parcoured_path_bonus.c	\
 									insert_path_node_bonus.c	\
 									pathfinding_bonus.c			\
-									free_stack_bonus.c			\
-									lock_tiles_bonus.c
+									t_pathfinding_free_bonus.c	\
+									lock_tiles_bonus.c			\
+									add_neighboring_tiles_bonus.c
 PATHFINDING_BONUS				=	$(addprefix $(PATHFINDING_DIR_BONUS), $(PATHFINDING_FILES_BONUS))
 
 # UTILS
